# Avoid body parsing for requests with no Content-Type and no body