--- a/lib/handleRequest.js
+++ b/lib/handleRequest.js
@@ -11,18 +11,27 @@ const inputSchemaError = fastJsonStringify(schemas.inputSchemaError)
 
 function handleRequest (req, res, params, context) {
   var method = req.method
-  var request = new context.Request(params, req, urlUtil.parse(req.url, true).query, req.headers, req.log)
+  var headers = req.headers
+  var request = new context.Request(params, req, urlUtil.parse(req.url, true).query, headers, req.log)
   var reply = new context.Reply(res, context, request)
 
   if (method === 'GET' || method === 'HEAD') {
     return handler(reply)
   }
 
-  var contentType = req.headers['content-type']
+  var contentType = headers['content-type']
 
   if (method === 'POST' || method === 'PUT' || method === 'PATCH') {
-    // application/json content type
-    if (contentType && contentType.indexOf('application/json') > -1) {
+    if (contentType === undefined) {
+      if (
+        headers['transfer-encoding'] === undefined &&
+        (headers['content-length'] === '0' || headers['content-length'] === undefined)
+      ) { // Request has no body to parse
+        handler(reply)
+        return
+      }
+      // Fall through to check for a custom parser
+    } else if (contentType.indexOf('application/json') > -1) {
       return jsonBody(request, reply, context._jsonParserOptions)
     }