Several issues have been opened in the past because by default, Fastify responds with an error to POST requests without a body. People expect requests without a body to work without having to set a "catch all" content type parser. It's counter-intuitive to have to set a body parser to be able to handle requests without a body.

This change avoids running any content type parser if the request has no `Content-Type` header and no body. From [RFC 7230 Section 3.3](https://tools.ietf.org/html/rfc7230#section-3.3), we can determine if a request has a body by checking the `Transfer-Encoding` and `Content-Type` headers.

> The presence of a message body in a request is signaled by a Content-Length or Transfer-Encoding header field.

The rest of that section does use "SHOULD" a lot, but it looks like most tools (including Node, curl, and the browser's `fetch()` API) automatically set one of those headers. If someone ever sends a request that has a body but no `Content-Length` header, no `Transfer-Encoding` header, *and* no `Content-Type` header, it seems fair to tell them that they *should* set one of those headers (as opposed to giving an error to people who have done nothing wrong by sending a request without a body).

Fixes #297, #450, #537
This would have helped with #419 as well

<details>
<summary>Benchmarks</summary>

**master**
```
[1] Stat         Avg     Stdev   Max
[1] Latency (ms) 3.1     10.95   361
[1] Req/Sec      31422.4 6177.53 35647
[1] Bytes/Sec    4.71 MB 949 kB  5.51 MB
[1]
[1] 157k requests in 5s, 23.4 MB read
...
[1] Stat         Avg     Stdev   Max
[1] Latency (ms) 2.99    10.71   372
[1] Req/Sec      32587.2 6824.94 37791
[1] Bytes/Sec    4.86 MB 1.04 MB 5.77 MB
[1]
[1] 163k requests in 5s, 24.3 MB read
```

**this PR**
```
[1] Stat         Avg     Stdev   Max
[1] Latency (ms) 2.93    10.44   251
[1] Req/Sec      33342.4 6996.48 38751
[1] Bytes/Sec    4.99 MB 1.04 MB 6.03 MB
[1]
[1] 167k requests in 5s, 24.8 MB read
...
[1] Stat         Avg     Stdev   Max
[1] Latency (ms) 3.03    10.85   392
[1] Req/Sec      32244.8 6703.57 37183
[1] Bytes/Sec    4.81 MB 1.01 MB 5.77 MB
[1]
[1] 161k requests in 5s, 24 MB read
```
</details>

#### Checklist

- [x] run `npm run test` and `npm run benchmark`
- [x] tests and/or benchmarks are included
- [ ] documentation is changed or added
- [x] commit message and code follows [Code of conduct](https://github.com/fastify/fastify/blob/master/CODE_OF_CONDUCT.md)