@allevo 2018-01-12
What happen with chunked request that hasn't a content length? Has set the transfer encoding header?