@nwoltman 2018-01-12
@allevo Yes. For those requests the `Transfer-Encoding: chunked` header is set. There's already a test that covers this case:
https://github.com/fastify/fastify/blob/e3b8f8e3d1d547f2569da9d4528d0a904c94011b/test/custom-parser.test.js#L393-L429