@mcollina 2018-01-12
I'm not 100% sure this is a breaking change, but rather a bugfix.