@eljefedelrodeodeljefe 2018-01-23
Sorry, if I missed parts of the discussion, but imo this should include `Content-Type: application/json` with no body if it doesn't