@nwoltman 2018-01-23
@eljefedelrodeodeljefe I don't understand what you're suggesting. Could you elaborate?